Allow rematerializing CIncOffset+CSetBounds for stack vars
This allows us to massively reduce the number of stack spills at the cost
of (potentially) a few added instructions. Adding a new Pseudo tricks
LLVM into thinking that a bounded stack cap is a single instructions.
Without the pseudo I could not find a way to tell it to rematerialize.